requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.27.1
pandas==1.4.0
pendulum==2.1.2
azure-storage-file-datalake==12.5.0
azure-cosmos==4.2.0

